getTimestamp() has type number, but returns java.lang.Long #10146 (#11796)

+++ modules/lib/lib-node/src/main/java/com/enonic/xp/lib/node/GetNodeVersionsResultMapper.java
@@ -11,7 +11,7 @@ public class GetNodeVersionsResultMapper
 {
     private final NodeVersions nodeVersions;
 
-    private final long count;
+    private final int count;
 
     private final long total;
